+++ src/behaviours.js
@@ -1,8 +1,8 @@
-import Core from "css-modules-loader-core";
+import Core from "./css-loader-core";
 
 export const behaviours = {
   LOCAL: "local",
-  GLOBAL: "global"
+  GLOBAL: "global",
 };
 
 export function getDefaultPlugins(behaviour, generateScopedName) {
@@ -13,10 +13,10 @@ export function getDefaultPlugins(behaviour, generateScopedName) {
       Core.values,
       Core.localByDefault,
       Core.extractImports,
-      scope
+      scope,
     ],
 
-    [behaviours.GLOBAL]: [Core.values, Core.extractImports, scope]
+    [behaviours.GLOBAL]: [Core.values, Core.extractImports, scope],
   };
 
   return plugins[behaviour];
@@ -25,7 +25,7 @@ export function getDefaultPlugins(behaviour, generateScopedName) {
 export function isValidBehaviour(behaviour) {
   return (
     Object.keys(behaviours)
-      .map(key => behaviours[key])
+      .map((key) => behaviours[key])
       .indexOf(behaviour) > -1
   );
 }