+++ src/generateScopedName.js
@@ -3,9 +3,7 @@ import stringHash from "string-hash";
 export default function generateScopedName(name, filename, css) {
   const i = css.indexOf(`.${name}`);
   const lineNumber = css.substr(0, i).split(/[\r\n]/).length;
-  const hash = stringHash(css)
-    .toString(36)
-    .substr(0, 5);
+  const hash = stringHash(css).toString(36).substr(0, 5);
 
   return `_${name}_${hash}_${lineNumber}`;
 }